+++ library/src/schemas/object/object.ts
@@ -162,7 +162,7 @@ export function object<TObjectShape extends ObjectShape>(
           }
 
           // Otherwise, add value to object
-        } else {
+        } else if (result.output !== undefined || key in input) {
           output[key] = result.output;
         }
       }