+++ library/src/schemas/object/objectAsync.ts
@@ -171,7 +171,7 @@ export function objectAsync<TObjectShape extends ObjectShapeAsync>(
                 }
 
                 // Otherwise, add value to object
-              } else {
+              } else if (result.output !== undefined || key in input) {
                 output[key] = result.output;
               }
             }