+++ library/src/transformations/toLowerCase/toLowerCase.ts
@@ -1,4 +1,5 @@
 import type { PipeResult } from '../../types.ts';
+import { getOutput } from '../../utils/index.ts';
 
 /**
  * Creates a transformation function that converts all the alphabetic
@@ -7,7 +8,6 @@ import type { PipeResult } from '../../types.ts';
  * @returns A transformation function.
  */
 export function toLowerCase() {
-  return (input: string): PipeResult<string> => ({
-    output: input.toLocaleLowerCase(),
-  });
+  return (input: string): PipeResult<string> =>
+    getOutput(input.toLocaleLowerCase());
 }