+++ library/src/transformations/toTrimmedStart/toTrimmedStart.ts
@@ -1,4 +1,5 @@
 import type { PipeResult } from '../../types.ts';
+import { getOutput } from '../../utils/index.ts';
 
 /**
  * Creates a transformation function that removes the leading white space and
@@ -7,7 +8,5 @@ import type { PipeResult } from '../../types.ts';
  * @returns A transformation function.
  */
 export function toTrimmedStart() {
-  return (input: string): PipeResult<string> => ({
-    output: input.trimStart(),
-  });
+  return (input: string): PipeResult<string> => getOutput(input.trimStart());
 }