+++ library/src/validations/equal/equal.ts
@@ -4,6 +4,8 @@ import { getOutput, getPipeIssues } from '../../utils/index.ts';
 /**
  * Creates a validation function that checks the value for equality.
  *
+ * @deprecated Function has been renamed to `value`.
+ *
  * @param requirement The required value.
  * @param error The error message.
  *