+++ library/src/validations/isoTime/isoTime.ts
@@ -2,7 +2,7 @@ import type { ErrorMessage, PipeResult } from '../../types.ts';
 import { getOutput, getPipeIssues } from '../../utils/index.ts';
 
 /**
- * Creates a validation functions that validates a time.
+ * Creates a validation function that validates a time.
  *
  * Format: hh:mm
  *